feat(babel): replace plain IEntrypoint object with class
