+++ packages/babel/src/index.ts
@@ -17,10 +17,10 @@ export { default as withLinariaMetadata } from './utils/withLinariaMetadata';
 export { default as Module, DefaultModuleImplementation } from './module';
 export { default as transform } from './transform';
 export * from './types';
-export { parseFile } from './transform-stages/helpers/parseFile';
+export { parseFile } from './transform-stages/queue/Entrypoint.helpers';
 export { default as loadLinariaOptions } from './transform-stages/helpers/loadLinariaOptions';
 export { prepareCode } from './transform-stages/queue/generators/transform';
-export { createEntrypoint } from './transform-stages/queue/createEntrypoint';
+export { Entrypoint } from './transform-stages/queue/Entrypoint';
 export { transformUrl } from './transform-stages/queue/generators/extract';
 export { default as isNode } from './utils/isNode';
 export { default as getTagProcessor } from './utils/getTagProcessor';
